Use insta for exporter tests
