+++ Cargo.toml
@@ -63,8 +63,13 @@ features = [
 [dev-dependencies]
 approx = "0.5"
 assert_cmd = "2.0"
+insta = { version = "1.41.1", features = ["yaml"] }
 predicates = "3.0"
 tempfile = "3.14"
+
+[profile.dev.package]
+insta.opt-level = 3
+similar.opt-level = 3
 
 [build-dependencies]
 clap = "4.4.12"
